final-pjt-back/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.16.0
environ==1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
openai==1.59.7
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.8.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.121
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
Wikipedia-API==0.7.3
youtube-transcript-api==1.0.3
zipp==3.21.0
